+++ src/directives/searchbox.js
@@ -19,7 +19,7 @@
             restrict: 'E',
             replace: true,
             template: '<div class="odswidget odswidget-searchbox">' +
-                    '<form method="GET" action="{{ actionUrl }}">' +
+                    '<form method="GET" action="{{ actionUrl }}" ng-if="actionUrl">' +
                     '<input class="searchbox" name="q" type="text" placeholder="{{placeholder}}">' +
                     '</form>' +
                 '</div>',